moved deployment code to src folder
