+++ package.json
@@ -1,11 +1,11 @@
 {
-    "name": "basic-auth",
+    "name": "riverserver",
     "version": "1.0.0",
-    "description": "A basic passport authentication boilerplate",
-    "main": "server.js",
+    "description": "Server for the RiverWiki - handles authentication and guide posts",
+    "main": "scr/server.js",
     "scripts": {
-        "start": "node server.js",
-        "dev": "nodemon server.js",
+        "start": "node src/server.js",
+        "dev": "nodemon src/server.js",
         "test": "jest"
     },
     "keywords": [
