+++ src/validation/is-empty.js
